Moving out some utility code
